Refactor app into helper modules
